@@ -6,29 +6,29 @@
   "repository": "github.com/gobuffalo/buffalo",
   "dependencies": {
     "@babel/preset-react": "^7.0.0",
-    "bootstrap": "3",
-    "bootstrap-sass": "^3.4.0",
+    "bootstrap": "^3.4.1",
+    "bootstrap-sass": "^3.4.1",
     "bounce": "^1.2.3",
     "bounce.js": "^0.8.2",
     "font-awesome": "~4.7.0",
-    "jquery": "~3.2.1",
+    "jquery": "~3.3.1",
     "jquery-ujs": "~1.2.2",
     "monaco-editor-webpack-plugin": "^1.7.0",
     "noty": "^3.2.0-beta",
-    "popper.js": "^1.14.4",
-    "react": "^16.7.0-alpha.0",
+    "popper.js": "^1.14.7",
+    "react": "^16.8.2",
     "react-bootstrap": "^0.32.4",
-    "react-dom": "^16.7.0-alpha.0",
+    "react-dom": "^16.8.2",
     "react-event-timeline": "^1.6.3",
     "react-monaco-editor": "^0.22.0",
     "react-router-dom": "^4.3.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.3",
+    "@babel/preset-env": "^7.3.1",
     "@babel/preset-es2015": "^7.0.0-beta.53",
-    "babel-loader": "^8.0.0-beta.6",
+    "babel-loader": "^8.0.5",
     "babel-preset-es2015": "^6.24.1",
     "copy-webpack-plugin": "~4.5.2",
     "css-loader": "~1.0.0",
